'Divide' of the database class
