+++ SuperGamblino/Commands/ShowProfile.cs
@@ -1,6 +1,7 @@
 ﻿using System.Threading.Tasks;
 using DSharpPlus.CommandsNext;
 using DSharpPlus.CommandsNext.Attributes;
+using SuperGamblino.DatabaseConnectors;
 using SuperGamblino.GameObjects;
 using static SuperGamblino.GameObjects.Work;
 
@@ -8,13 +9,13 @@ namespace SuperGamblino.Commands
 {
     internal class ShowProfile
     {
-        private readonly Database _database;
+        private readonly UsersConnector _usersConnector;
         private readonly Messages _messages;
 
-        public ShowProfile(Database database, Messages messages)
+        public ShowProfile(Messages messages, UsersConnector usersConnector)
         {
-            _database = database;
             _messages = messages;
+            _usersConnector = usersConnector;
         }
 
         [Command("profile")]
@@ -22,7 +23,7 @@ namespace SuperGamblino.Commands
         [Description("Shows a profile, descriping detailed information about the user. This command has no arguments.")]
         public async Task OnExecute(CommandContext command)
         {
-            User user = await _database.GetUser(command);
+            User user = await _usersConnector.GetUser(command.User.Id);
             Job currentJob = GetCurrentJob(user.Level);
             await _messages.Profile(command, user, currentJob);
         }
