+++ SuperGamblino/GameObjects/User.cs
@@ -12,6 +12,5 @@ namespace SuperGamblino.GameObjects
         public DateTime? LastDailyReward { get; set; }
         public int Experience { get; set; }
         public int Level { get; set; }
-        public DiscordUser DiscordUser { get; set; }
     }
 }
