+++ SuperGamblino/Helpers/Exp.cs
@@ -4,26 +4,27 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
+using SuperGamblino.DatabaseConnectors;
 
 namespace SuperGamblino.Helpers
 {
     class Exp
     {
-        private readonly Database _database;
+        private readonly UsersConnector _usersConnector;
 
-        public Exp (Database database)
+        public Exp (UsersConnector usersConnector)
         {
-            _database = database;
+            _usersConnector = usersConnector;
         }
         public async Task<AddExpResult> Give(CommandContext command, int bet)
         {
-            User user = await _database.GetUser(command);
+            User user = await _usersConnector.GetUser(command.User.Id);
             Random rnd = new Random();
             int exp = rnd.Next(50, 125);
             if (user.Level * 15 >= bet)
-                return await _database.CommandGiveUserExp(command, 0);
+                return await _usersConnector.CommandGiveUserExp(command, 0);
             else
-                return await _database.CommandGiveUserExp(command, exp);
+                return await _usersConnector.CommandGiveUserExp(command, exp);
         }
     }
 }
